move test for functions in utils.py into a separate test file
